requirements.txt
Flask
Gunicorn
python-dotenv
pymongo[srv]
Flask-PyMongo
google-generativeai
Flask-Cors
langchain
google
PyJWT
pinecone
openai
langchain-community
langchain-community
langchain_openai
langchain_core
langchain_core
redis
docx2txt
pandas
openpyxl



langchain-openai==0.0.5
langchain-community==0.0.13